Suggested changes applied
