+++ lib/cmd_n2ip.go
@@ -61,9 +61,8 @@ func CmdN2IP(f CmdN2IPFlags, args []string, printHelp func()) error {
 		return err
 	}
 
-	postfix := InfixToPostfix(tokens)
-
 	// Evaluate the postfix expression
+	postfix := InfixToPostfix(tokens)
 	result, err := EvaluatePostfix(postfix)
 	if err != nil {
 		return err
